requirements.txt
numpy
tensorflow-gpu==1.12.0
matplotlib
scipy
imageio

